:truck: Rename `capabilities` -> `server_capabilities`
